+++ api/tests/views/__init__.py
@@ -9,4 +9,7 @@ from api.tests.views.history import *
 from api.tests.views.webhook import *
 from api.tests.views.bundle import *
 from api.tests.views.glossary import *
-from api.tests.views.mcp_expansion import *
+from api.tests.views.mcp.test_protocol import *
+from api.tests.views.mcp.test_tools_project import *
+from api.tests.views.mcp.test_tools_tokens import *
+from api.tests.views.mcp.test_tools_ai import *
