+++ api/views/mcp/__init__.py
@@ -1,0 +1,3 @@
+from .view import McpView
+
+__all__ = ["McpView"]
