+++ api/views/mcp/tools_project.py
@@ -1,0 +1,12 @@
+from api.models.language import Language
+
+
+def get_project(args, access):
+    p = access.project
+    return {"id": p.id, "name": p.name, "description": p.description}
+
+
+def get_languages(args, access):
+    codes = list(Language.objects.filter(
+        project=access.project).values_list('code', flat=True))
+    return {"languages": codes}
